+++ src/api/java/com/codetaylor/mc/artisanworktables/api/recipe/IArtisanRecipe.java
@@ -179,6 +179,8 @@ public interface IArtisanRecipe {
    */
   int getToolCount();
 
+  ToolEntry[] getToolEntries();
+
   /**
    * @return an immutable list of the recipe's ingredients
    */